Tweak to scale for new unified

+++ src/PROcess.cxx
@@ -67,7 +67,7 @@ namespace PROfit {
         if(binned) {
             Eigen::VectorXf systw = Eigen::VectorXf::Constant(inconfig.m_num_variable_bins_total[var_index], 1);
             for(int i = 0; i < shifts.size(); ++i) {
-                int binning = insyst.spline_binnings[i];
+                size_t binning = insyst.spline_binnings[i];
                 const Eigen::MatrixXf &hist = inprop.variable_hist_storage(binning,var_index);
                 for(size_t k = 0; k < inconfig.m_num_variable_bins_total[var_index]; ++k) {
                     if(binning == var_index){
@@ -251,7 +251,7 @@ namespace PROfit {
         float spline_throw = d(rng);
         int binning = insyst.spline_binnings[spline];
 
-        if(other_index == inconfig.i_prime) {
+        if(other_index == (int)inconfig.i_prime) {
             const Eigen::MatrixXf &hist = inprop.variable_hist_storage(binning,other_index);
             for(long int i = 0; i < hist.rows(); ++i) {
                 float systw = 1.0;
